(incomplete, ammend later) Add LoadConfig functionality
This commit should load the configurations from a simple txt file
and set them as CURL options.

For now, only the record to trigger it is added.

Initially I had used a toml file, but I'm starting to think using
a custom parser would be better just for sake of simplicity.